fix(#900): fix SU history species minimum height type

+++ backend/src/main/java/ca/bc/gov/restapi/results/oracle/dto/opening/history/OpeningStandardUnitHistorySpeciesDetailsDto.java
@@ -22,8 +22,8 @@ public record OpeningStandardUnitHistorySpeciesDetailsDto(
     CodeDescriptionDto newSpecies,
 
     @Schema(types = {"number", "null"}, requiredMode = Schema.RequiredMode.REQUIRED)
-    Integer oldMinHeight,
+    Double oldMinHeight,
 
     @Schema(types = {"number", "null"}, requiredMode = Schema.RequiredMode.REQUIRED)
-    Integer newMinHeight
+    Double newMinHeight
 ) {}
